Moves method to MoveValidator class
